M2Crypto/SSL/TwistedProtocolWrapper.py : Fix heartbleed checker issue
If an heartbleed checker pass on any servder user TwistedProtocolWrapper from
M2Crypto the process turns crazy and never exit from an infinite loop.

In this case it appends when the data is not "None" but is empty string.
I juste test if the data is empty and break to exit the loop.